Fetch emails from NNTP instead of IMAP

+++ src/Application/Command/DbCommand.php
@@ -59,11 +59,16 @@ class DbCommand
     public function purge(bool $force, OutputInterface $output)
     {
         $tables = $this->db->getSchemaManager()->listTableNames();
-        foreach ($tables as $table) {
-            $output->writeln("<info>Dropping table $table</info>");
-            if ($force) {
-                $this->db->getSchemaManager()->dropTable($table);
+
+        if ($force) $this->db->exec('SET FOREIGN_KEY_CHECKS=0');
+
+        try {
+            foreach ($tables as $table) {
+                $output->writeln("<info>Dropping table $table</info>");
+                if ($force) $this->db->getSchemaManager()->dropTable($table);
             }
+        } finally {
+            if ($force) $this->db->exec('SET FOREIGN_KEY_CHECKS=1');
         }
 
         if (!$force) {
@@ -73,5 +78,27 @@ class DbCommand
         }
 
         $this->setup($force, $output);
+    }
+
+    public function truncate(bool $force, OutputInterface $output)
+    {
+        $tables = $this->db->getSchemaManager()->listTableNames();
+
+        if ($force) $this->db->exec('SET FOREIGN_KEY_CHECKS=0');
+
+        try {
+            foreach ($tables as $table) {
+                $output->writeln("<info>Truncating table $table</info>");
+                if ($force) $this->db->exec("TRUNCATE TABLE $table");
+            }
+        } finally {
+            if ($force) $this->db->exec('SET FOREIGN_KEY_CHECKS=1');
+        }
+
+        if (!$force) {
+            $output->writeln('<comment>No query was run, use the --force option to run the queries</comment>');
+        } else {
+            $output->writeln('<comment>Queries were successfully run against the database</comment>');
+        }
     }
 }
